shared build of glew and logging (#145)
* build easylogging++ into util module

* fix a typo

* fix  a warning

* use float for duration

* build glew into renderer

* shared library

* clean windows

* fix several warnings

* update

* clean

* opengl lib name

* install

* handle Qt version

* link to OpenGL libraries

* Update cmake-gcc.yml (#144)

+++ easy3d/renderer/camera.h
@@ -254,7 +254,7 @@ namespace easy3d {
         /// Perform keyframe interpolation to fit the scene bounding box.
         void interpolateToFitScene();
         /// Perform keyframe interpolation to \p frame, done in \p duration second.
-        void interpolateTo(const Frame &frame, double duration);
+        void interpolateTo(const Frame &frame, float duration);
 
 	public:
 		/**
